+++ engine/runtime-instrument-common/src/main/scala/org/enso/interpreter/instrument/command/CloseFileCmd.scala
@@ -3,28 +3,25 @@ package org.enso.interpreter.instrument.command
 import org.enso.interpreter.instrument.execution.RuntimeContext
 import org.enso.polyglot.runtime.Runtime.Api
 
-import scala.concurrent.{ExecutionContext, Future}
-
 /** A command that closes a file.
   *
   * @param request a request for a service
   */
-class CloseFileCmd(request: Api.CloseFileNotification) extends Command(None) {
+class CloseFileCmd(request: Api.CloseFileNotification)
+    extends SynchronousCommand(None) {
 
-  /** @inheritdoc */
-  override def execute(implicit
-    ctx: RuntimeContext,
-    ec: ExecutionContext
-  ): Future[Unit] =
-    Future {
-      ctx.locking.acquireReadCompilationLock()
-      ctx.locking.acquireFileLock(request.path)
-      try {
-        ctx.executionService.resetModuleSources(request.path)
-      } finally {
-        ctx.locking.releaseFileLock(request.path)
-        ctx.locking.releaseReadCompilationLock()
-      }
+  override def executeSynchronously(implicit ctx: RuntimeContext): Unit = {
+    ctx.locking.acquireReadCompilationLock()
+    ctx.locking.acquireFileLock(request.path)
+    ctx.locking.acquirePendingEditsLock()
+    try {
+      ctx.state.pendingEdits.dequeue(request.path)
+      ctx.executionService.resetModuleSources(request.path)
+    } finally {
+      ctx.locking.releasePendingEditsLock()
+      ctx.locking.releaseFileLock(request.path)
+      ctx.locking.releaseReadCompilationLock()
     }
+  }
 
 }