+++ engine/runtime-instrument-common/src/main/scala/org/enso/interpreter/instrument/command/StartBackgroundProcessingCmd.scala
@@ -12,10 +12,10 @@ import scala.concurrent.{ExecutionContext, Future}
   */
 final class StartBackgroundProcessingCmd(
   maybeRequestId: Option[Api.RequestId]
-) extends Command(maybeRequestId) {
+) extends AsynchronousCommand(maybeRequestId) {
 
   /** @inheritdoc */
-  override def execute(implicit
+  override def executeAsynchronously(implicit
     ctx: RuntimeContext,
     ec: ExecutionContext
   ): Future[Unit] = {